requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0